+++ src/parser/error.rs
@@ -1,4 +1,9 @@
-use std::fmt::{Display, Write};
+use std::fmt::Display;
+
+use winnow::{
+    error::{AddContext, ParserError},
+    stream::{LocatingSlice, Offset},
+};
 
 use crate::{error::LocatedError, parser::LSpan};
 
@@ -16,9 +21,11 @@ pub enum ErrorKind {
 /// Error type for [parse][super::parse].
 #[derive(Debug, Clone, PartialEq, Eq)]
 pub struct Error<'a> {
-    nom_kind: nom::error::ErrorKind,
-    pub(crate) span: LSpan<'a>,
     kind: Option<ErrorKind>,
+    // The remaining input at the error point. Used for display of the first character.
+    pub(crate) input: LSpan<'a>,
+    // Line number in source
+    pub(crate) line: usize,
 }
 
 impl Display for Error<'_> {
@@ -32,34 +39,44 @@ impl Display for Error<'_> {
             None => "unknown error",
         })?;
 
-        f.write_char(' ')?;
-        if let Some(c) = self.span.chars().next() {
-            f.write_char('\'')?;
-            f.write_char(c)?;
-            f.write_char('\'')?;
-            f.write_char(' ')?;
+        if let Some(c) = self.input.chars().next() {
+            f.write_fmt(format_args!(" \'{}\'", c))?;
         }
 
-        f.write_fmt(format_args!(
-            "near {}:{}",
-            self.span.extra,
-            self.span.get_column()
-        ))
+        let (line, offset) = self.location();
+        f.write_fmt(format_args!(" near {}:{}", line, offset))
     }
 }
 
 impl std::error::Error for Error<'_> {}
 
-impl Error<'_> {
-    pub(crate) fn attach(mut self, kind: ErrorKind) -> Self {
-        if self.kind.is_none() {
-            self.kind = Some(kind);
+impl<'a> Error<'a> {
+    fn new(input: &LocatingSlice<&'a str>) -> Self {
+        Self {
+            kind: None,
+            input: *input,
+            line: 0,
         }
+    }
+
+    pub(crate) fn with_input(mut self, input: &LocatingSlice<&'a str>) -> Self {
+        self.input = *input;
+        self
+    }
+
+    /// Set the line number for this error.
+    pub(crate) fn with_line(mut self, line: usize) -> Self {
+        self.line = line;
+        self
+    }
+
+    /// Set the line number for this error.
+    pub(crate) fn with_kind(mut self, kind: ErrorKind) -> Self {
+        self.kind = Some(kind);
         self
     }
 
     /// Get error kind.
-    #[must_use]
     pub fn kind(&self) -> Option<ErrorKind> {
         self.kind
     }
@@ -67,20 +84,29 @@ impl Error<'_> {
 
 impl LocatedError for Error<'_> {
     fn location(&self) -> (usize, usize) {
-        (self.span.extra, self.span.get_column())
+        let mut start = self.input.clone();
+        start.reset_to_start();
+        (self.line + 1, self.input.offset_from(&start) + 1)
     }
 }
 
-impl<'a> nom::error::ParseError<LSpan<'a>> for Error<'a> {
-    fn from_error_kind(input: LSpan<'a>, kind: nom::error::ErrorKind) -> Self {
-        Self {
-            nom_kind: kind,
-            span: input,
-            kind: None,
-        }
+impl<'a> ParserError<LSpan<'a>> for Error<'a> {
+    type Inner = Self;
+
+    fn from_input(input: &LocatingSlice<&'a str>) -> Self {
+        Self::new(input)
     }
 
-    fn append(_input: LSpan<'a>, _kind: nom::error::ErrorKind, other: Self) -> Self {
-        other
+    fn into_inner(self) -> Result<Self::Inner, Self> {
+        Ok(self)
+    }
+}
+
+impl<'a> AddContext<LSpan<'a>, ErrorKind> for Error<'a> {
+    fn add_context(
+        self, _input: &LSpan<'a>, _token_start: &<LSpan<'a> as winnow::stream::Stream>::Checkpoint,
+        context: ErrorKind,
+    ) -> Self {
+        self.with_kind(context)
     }
 }