+++ src/parser/item.rs
@@ -1,21 +1,18 @@
-use crate::{
-    generator::{Tag, TagG},
-    parser::LSpan,
-};
+use crate::generator::{Tag, TagG};
 
 /// AST item.
 ///
 /// In parsing stage, each line of source code will be parsed as a `Vec<Item>`, so the final result
 /// is `Vec<Vec<Item>>`.
 ///
-/// Tag conversion stage, Each tag will be converted from [`LSpan`] into [Tag] type the generator
+/// Tag conversion stage, Each tag will be converted from [`&str`] into [Tag] type the generator
 /// needed, by using [`TagConvertor`][crate::generator::TagConvertor] of the generator.
 ///
-/// In generating stage, generator will convert `Vec<Vec<Item<'_, Tag>>>>` to final output.
+/// In generating stage, generator will convert `Vec<Vec<Item<'_, Tag>>>` to final output.
 #[derive(Debug, Clone, PartialEq)]
-pub enum Item<'a, Tag = LSpan<'a>> {
+pub enum Item<'a, Tag = &'a str> {
     /// Plain text(escaped) without any style.
-    PlainText(LSpan<'a>),
+    PlainText(&'a str),
     /// A styled element, contains a series tag name and subitems.
     Element(Vec<Tag>, Vec<Item<'a, Tag>>),
 }