+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- Migrated parser from `nom` / `nom-locate` to `winnow`.
+
 ## [0.5.0] - 2024-10-04
 
 ### Changed